multer

+++ client/src/components/Navbar.jsx
@@ -1,29 +1,39 @@
-import {Link ,useNavigate} from 'react-router-dom'
+// src/components/Navbar.jsx
+import { Link, useNavigate } from 'react-router-dom';
 
-export default function Navbar() {
-  const token = localStorage.getItem('token')
-  const navigate = useNavigate()
-  function logout(){
-    localStorage.clear()
-    navigate('/login')
-  }
-  function login(){
-    navigate('/login')
-  }
-  return (
-    <nav style={{display:"flex",justifyContent:'space-between', alignItems:'center',padding:"16px"}}>
-      <div style={{display:"flex", gap:"16px", }}>
-        <Link to={'/'}>Accueil</Link>
-        <Link to={'/cv'}>CV</Link>
-        <Link to={'/contact'}>Contact</Link>
-        {/* <Link to={'/login'}>Connexion</Link> */}
-        {/* <Link to={'/register'}>Inscription</Link>
-        <Link to={'/admin'}>dashboard</Link> */}
-      </div>
-      {token
-        ?<button onClick={logout}>Déconnexion</button>
-        :<button onClick={login}>Connexion</button>
-      }
-    </nav>
-  )
+export default function Navbar() { // On récupère user et setUser ici
+	const navigate = useNavigate();
+
+	async function handleLogout() {
+		try {
+			const response = await fetch('http://localhost:3000/logout', {
+				method: 'POST',
+				credentials: "include"
+			});
+
+			if (response.ok) {
+				localStorage.clear();
+				setUser(null); // Changement immédiat de l'affichage !
+				navigate('/login');
+			}
+		} catch (error) {
+			console.error("Erreur déconnexion :", error);
+		}
+	}
+
+	return (
+		<nav style={{ display: "flex", justifyContent: 'space-between', alignItems: 'center', padding: "16px" }}>
+			<div style={{ display: "flex", gap: "16px" }}>
+				<Link to={'/'}>Accueil</Link>
+
+
+				<Link to={'/login'}>Connexion</Link>
+				<Link to={'/register'}>Inscription</Link>
+
+				<Link to={'/admin'}>dashboard</Link>
+			</div>
+
+
+		</nav>
+	);
 }
